+++ src/index.ts
@@ -1,7 +1,7 @@
-export { default as crawlFavicons } from './crawlFavicons';
-export { default as crawlWithFetch } from './crawlWithFetch';
-export { default as crawlWithPuppeteer } from './crawlWithPuppeteer';
+export { default as fetchStringNode } from './fetchStringNode';
+export { default as fetchStringPuppeteer } from './fetchStringPuppeteer';
 export { default as findFavicons } from './findFavicons';
+export { default as getFavicons } from './getFavicons';
 export * from './constants';
 export * from './lib';
 export * from './types';
